backend/requirements.txt
fastapi[standard]
uvicorn[standard]
pydantic
python-dotenv
ddgs
tldextract
httpx
beautifulsoup4
lxml
google-genai
